logica implementada (quase tudo talvez?), falta os drivers

+++ main.c/conexao.h
@@ -33,7 +33,7 @@ extern STRUCT_CONN stConn;
 
 /* Funcoes */
 void vCONN_DescartaMensagem(STRUCT_CONN* struct_conn);
-void vCONN_LeituraPesos(uint8_t* pesos);
+void vCONN_LeituraPesos(uint16_t* pesos);
 void vCONN_FechaConexao();
 
 bool bCONN_SendAck();
