+++ main.c/conexao.ino
@@ -9,7 +9,7 @@ STRUCT_CONN stConn =
 };
 
 /* Função que pega os pesos do pacote bluetooth e põe no vetor pesos */
-void vCONNLeituraPesos(uint8_t* pesos)
+void vCONNLeituraPesos(uint16_t* pesos)
 {
 
   
@@ -21,16 +21,19 @@ void vCONN_Poll()
   
 }
 
+/* Envia um pacote de ACK para o aplicativo */
 bool bCONN_SendAck()
 {
   
 }
 
+/* Envia um pacote de informações de crédito ao aplicativo */
 bool bCONN_SendUserCredit()
 {
   
 }
 
+/* Envia um pacote de requisição de ACK ao aplicativo */
 bool bCONN_SendAckRequest()
 {
